New function added to get random phrase
Created a new function "getRandomPhraseAsArray" that selects a random phrase from the previous "phrases" array and stores it in "randomPhrase".

Created a new variable called "phraseLength" that determines how many characters are in the randomly chosen phrase, including spaces.

Created a new variable that splits the chosen phrase into separate characters and stores in a new string.